requirements.txt
ipdb==0.13.9
python-dotenv==0.13.0
twitchio==2.3.0